requirements.txt
apache-beam==2.25.0
apache-beam[gcp]==2.25.0
# Imports GCS client needed for beam
google-apitools==0.5.31
google-cloud-bigquery==2.3.1
google-cloud-storage==1.32.0
# Specify resumable-media version to avoid dependency conflict
google-resumable-media===1.1.0
httpio==0.3.0
pyasn==1.6.1
requests==2.24.0
retry==0.9.2
schedule==0.6.0
